requirements.txt
Django>=1.9
Pillow>=4.0
gunicorn
lxml
requests
django-solo-grappelli>=1.1.2
unicodecsv
gitpython